refactor: Updated setupIntl()
